Flytt ned dvhbaseentitet
